Adding check for `created_at` property in groups.py

+++ congregate/migration/gitlab/groups.py
@@ -168,14 +168,21 @@ class GroupsClient(BaseClass):
                 elif not dry_run:
                     try:
                         group = resp.json()
-                        # The groups API doesn't have timestamps available currently. This property needs to be exposed in the API before uncommenting these lines
-                        # if get_timedelta(group["created_at"]) < self.config.max_asset_expiration_time:
-                        self.groups_api.delete_group(
-                            group["id"],
-                            self.config.destination_host,
-                            self.config.destination_token)
-                        # else:
-                        #     self.log("Ignoring %s. Group existed before %d hours" % (group["path_with_namespace"], self.config.max_asset_expiration_time))
+                        if group.get("created_at", None):
+                            if get_timedelta(group["created_at"]) < self.config.max_asset_expiration_time:
+                                self.groups_api.delete_group(
+                                    group["id"],
+                                    self.config.destination_host,
+                                    self.config.destination_token)
+                            else:
+                                self.log("Ignoring %s. Group existed before %d hours" % (group["path_with_namespace"], self.config.max_asset_expiration_time))
+                        else:
+                            # TODO: Remove this block when the `created_at` field is properly exposed in the groups API
+                            self.log.warn("Unable to find timestamp for group %s. Deleting group anyway." % group["path_with_namespace"])
+                            self.groups_api.delete_group(
+                                    group["id"],
+                                    self.config.destination_host,
+                                    self.config.destination_token)
 
                     except RequestException, e:
                         self.log.error(
